requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==4.2.15
gunicorn==20.1.0
psycopg2==2.9.9
sqlparse==0.5.1
whitenoise==6.5.0
djangorestframework